Rename package "sas" to "verification"
